Move per-game smoke test runners into game_engine/tests/
The scripting/ engine-core folder also held ~40 headless smoke runners. Split
out the ones that test *games* (not engine subsystems): move the nine
per-game runners — pong / pacman / invaders / breakout / ylos / ylos2 /
spawner / counter / world_scroll _runner_demo.rgr — into
gallery/game_engine/tests/, alongside the existing interp/ tests.

- Each runner's only local dependency, Import "./game_runtime.rgr", is
  repointed to ../scripting/game_runtime.rgr (engine core stays put).
- The .game.tsx fixtures stay in scripting/: the same file is a shared input to
  the .d.ts type graph (tsconfig), the watch/build scripts, and the AOT native
  games in ranger_games/. The runners already read them by repo-root path, so
  the move needs no fixture changes.
- Engine-subsystem smoke runners (audio, UI, physics, catalog, wasm, …) stay in
  scripting/ next to the runtime they exercise.
- Update references: package.json (engine:pong:runner, engine:spawner:runner),
  tests/game-runner.test.ts, tests/ts-to-ranger-host.test.ts, and the runner
  command examples in GAME_SCRIPTING.md / GAME_ENGINE_DESIGN.md.
- Docs/skills: new tests/README.md, plus README.md and scripting/AGENTS.md.

Verified by compiling all nine moved runners to ES6 and running the pong and
spawner runners headless — the fixture read-paths still resolve.
