+++ gallery/game_engine/docs/GAME_SCRIPTING.md
@@ -77,7 +77,7 @@ Run a game headless and dump a PNG (Node harness):
 
 ```bash
 RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr \
-  node bin/output.js -es6 ./gallery/game_engine/scripting/pong_runner_demo.rgr \
+  node bin/output.js -es6 ./gallery/game_engine/tests/pong_runner_demo.rgr \
   -d=./tests/.output -o=pong_runner_demo.js -nodecli
 node ./tests/.output/pong_runner_demo.js 300
 ffmpeg -f rawvideo -pixel_format rgba -video_size 480x270 \
@@ -241,5 +241,5 @@ function update(props) {
 ```
 
 Minimal example: [`world_scroll.game.tsx`](../scripting/world_scroll.game.tsx)
-(headless: [`world_scroll_runner_demo.rgr`](../scripting/world_scroll_runner_demo.rgr)).
+(headless: [`world_scroll_runner_demo.rgr`](../tests/world_scroll_runner_demo.rgr)).
 Interpreter perf notes: [`TS_ENGINE_OPTIMIZATION.md`](./TS_ENGINE_OPTIMIZATION.md).
