+++ gallery/game_engine/scripting/AGENTS.md
@@ -22,9 +22,20 @@ Everything in `gallery/game_engine/scripting/` is one of:
   (e.g. `game_runtime.rgr`, `game_audio.rgr`, `game_vocal_fx.rgr`), plus the
   native/WASM/sprite host runners.
 - **Ranger test fixtures** — loose `*.game.tsx` scripts used by headless
-  runners to exercise the interpreter (e.g. `vocal_fx_demo.game.tsx`).
-- **Headless smoke runners** — `*_runner_demo.rgr` / `*_selftest.rgr` that
-  compile to JS and are driven by `npm run engine:*` scripts.
+  runners to exercise the interpreter (e.g. `vocal_fx_demo.game.tsx`). These
+  stay here even for games whose smoke runner moved to `../tests/`, because the
+  same `.game.tsx` is also a shared input to the `.d.ts` type graph
+  (`tsconfig.json`, `engine.d.ts`), the watch/build scripts, and the AOT native
+  games in `../ranger_games/` — the runners read them by repo-root path.
+- **Engine-subsystem smoke runners** — `*_runner_demo.rgr` / `*_selftest.rgr`
+  for engine core (audio, UI, physics, catalog, wasm, background…) that compile
+  to JS and are driven by `npm run engine:*`. These exercise the *runtime*, so
+  they live next to it.
+
+> The **per-game** smoke runners (`pong`/`pacman`/`invaders`/`breakout`/`ylos`/
+> `ylos2`/`spawner`/`counter`/`world_scroll`_runner_demo.rgr) moved to
+> [`../tests/`](../tests/). Their `.game.tsx` fixtures stayed here (shared input,
+> see above); the runners read them via a repo-root path.
 
 **These are not playable games and they do not appear in the launcher menu.**
 The launcher (`menu/`, `game_catalog.rgr`) only scans **game folders**:
