+++ gallery/game_engine/tests/breakout_runner_demo.rgr
@@ -2,7 +2,7 @@
 ; breakout_runner_demo - runs breakout.game.tsx via GameRunner (multi-screen).
 ; ==============================================================================
 
-Import "./game_runtime.rgr"
+Import "../scripting/game_runtime.rgr"
 
 class BreakoutRunnerDemo {
     sfn m@(main):void () {
