+++ gallery/game_engine/tests/invaders_runner_demo.rgr
@@ -6,7 +6,7 @@
 ; next to this file for visual inspection.
 ; ==============================================================================
 
-Import "./game_runtime.rgr"
+Import "../scripting/game_runtime.rgr"
 
 class InvadersRunnerDemo {
     sfn m@(main):void () {
