+++ gallery/game_engine/tests/spawner_runner_demo.rgr
@@ -8,7 +8,7 @@
 ; spawner_native_runner.rgr. This proves the bridge contract is path-agnostic.
 ; ==============================================================================
 
-Import "./game_runtime.rgr"
+Import "../scripting/game_runtime.rgr"
 
 class SpawnerRunnerDemo {
     sfn m@(main):void () {
