+++ package.json
@@ -206,12 +206,12 @@
     "tsparser": "npm run tsparser:compile && npm run tsparser:run",
     "ts2ranger:compile": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/ts_to_ranger/ts_emitter_main.rgr -d=./gallery/ts_to_ranger/bin -o=ts_emitter_main.js -nodecli && node gallery/ts_to_ranger/scripts/patch-emitter-utf8.mjs",
     "ts2ranger:emit": "npm run ts2ranger:compile && node ./gallery/ts_to_ranger/bin/ts_emitter_main.js",
-    "engine:pong:runner": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/scripting/pong_runner_demo.rgr -d=./tests/.output -o=pong_runner_demo.js -nodecli && node ./tests/.output/pong_runner_demo.js 180",
+    "engine:pong:runner": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/tests/pong_runner_demo.rgr -d=./tests/.output -o=pong_runner_demo.js -nodecli && node ./tests/.output/pong_runner_demo.js 180",
     "engine:pong:native": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/ranger_games/pong_native_runner.rgr -d=./tests/.output -o=pong_native_runner.js -nodecli && node ./tests/.output/pong_native_runner.js 180",
     "engine:invaders:native": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/ranger_games/invaders_native_runner.rgr -d=./tests/.output -o=invaders_native_runner.js -nodecli && node ./tests/.output/invaders_native_runner.js 600",
     "engine:invaders:native:cpp": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -l=cpp ./gallery/game_engine/ranger_games/invaders_native_runner.rgr -d=./tests/.output -o=invaders_native_runner.cpp -nodecli && cp gallery/invaders/variant.hpp tests/.output/variant.hpp && g++ -std=c++17 -pthread tests/.output/invaders_native_runner.cpp -o tests/.output/invaders_native_runner_cpp.exe && ./tests/.output/invaders_native_runner_cpp.exe 600",
     "engine:pong:native:cpp": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -l=cpp ./gallery/game_engine/ranger_games/pong_native_runner.rgr -d=./tests/.output -o=pong_native_runner.cpp -nodecli && cp gallery/invaders/variant.hpp tests/.output/variant.hpp && g++ -std=c++17 -pthread tests/.output/pong_native_runner.cpp -o tests/.output/pong_native_runner_cpp.exe && ./tests/.output/pong_native_runner_cpp.exe 180",
-    "engine:spawner:runner": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/scripting/spawner_runner_demo.rgr -d=./tests/.output -o=spawner_runner_demo.js -nodecli && node ./tests/.output/spawner_runner_demo.js 30",
+    "engine:spawner:runner": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/tests/spawner_runner_demo.rgr -d=./tests/.output -o=spawner_runner_demo.js -nodecli && node ./tests/.output/spawner_runner_demo.js 30",
     "engine:spawner:native": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/ranger_games/spawner_native_runner.rgr -d=./tests/.output -o=spawner_native_runner.js -nodecli && node ./tests/.output/spawner_native_runner.js 30",
     "engine:vocalfx": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/scripting/vocal_fx_runner_demo.rgr -d=./tests/.output -o=vocal_fx_runner_demo.js -nodecli && node ./tests/.output/vocal_fx_runner_demo.js 170",
     "engine:vocalfx:selftest": "cross-env RANGER_LIB=./compiler/Lang.rgr:./lib/stdops.rgr node bin/output.js -es6 ./gallery/game_engine/scripting/vocal_fx_selftest.rgr -d=./tests/.output -o=vocal_fx_selftest.js -nodecli && node ./tests/.output/vocal_fx_selftest.js",
